implemented browser routing
