+++ package.json
@@ -11,6 +11,7 @@
     "@testing-library/user-event": "^13.5.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.3",
     "react-scripts": "^5.0.1",
     "use-local-storage": "^2.3.6",
     "web-vitals": "^2.1.4"
